Update parings_debug.py
Fixed an issue with the ratings being displayed incorrectly. The issue was that I was trying to generate ratings from the unsorted chart by using a sorted list of names and it was associating the wrong rows and columns. I fixed it by  moving the sorting of names to after the assignment of ratings.