Treeview color coding
added colors to the treeview. When the line has a value rating, it will map the rating and use the maximum of two different ratings to display a color.

Currently using the same color map for the treeview tags. 

Want to consider moving to a color heatmap in the future to support fractional matchup strength.