I think I fixed the math!
I am applying a max() function to the current set of child nodes.
This will return only the higher of the two options at each
point of choice for the user (or the opponent team).

The goal is to reduce the number volume. My thought is that the lower the numbers
the more constrained the calculations for the user to be making.