ADDED CYCLE BUTTON
Added a button to cycle through the scenarios.
currently this button will build a string to save as a csv.

Needs additional work.
This is currently going to overwrite the current work you've done on THIS scenario.
in a future version would just like to find a way to start at the current scenario and cycle from there.