fixed saving current values to csv
This will work so long as the csv is fully populated.

The grid needs to pull values from the csv to create it's rows each time.
If the text is off, it will throw off the grid during cycling.

Need a clean data set to cycle through each scenario and record all the results.
Thinking of exploring using this operation to recreate the text in the grid after the fact.