MATH BUTTON
I created a math button and mapped the traverse and sum values function to this button.

This allows us to keep the math functionality in the app temporarily
But it won't get called at the same time that the tree gets generated.

Hopefully this helps take some of the load off testing, we can trigger this extra load only if we want to.