Added multiple functions
added a new class to handle deleting teams with a pop up windown
features a drop down to pull teams from the db so it requried a class.

Added more functionality for import. Still not finished on import csv.
Wanted to pull the new managment functions in the dbmanager class.