Data rework (#12)
* wip

* in progress adding grid fill from db

* working prototype to load/save data to database

* removed textbox stuff to clean it up

* testing keys

* MATH BUTTON

Added math button to pull the match rating sums out of the tree generation.

* working on import csv

woo I managed to add the team name from the import file into the teams table (but only if it does not already exist in that table).

* progres on import

made progress on importing csv. You need to click a few times for some reason.
the first time imports the team name. After that it will import the player names for that team.
Still not done with the ratings imports.

* added an add team button

added a button and a function just to get team names into db.

* Added main.py

Added main.py to eventually support exe

* Worked on adding better import operations.

using chatgpt to generate the bulk of these operations. Still not finshed formating it to match our program exactly.

* wip on excel management

* wip

* Added multiple functions

added a new class to handle deleting teams with a pop up windown
features a drop down to pull teams from the db so it requried a class.

Added more functionality for import. Still not finished on import csv.
Wanted to pull the new managment functions in the dbmanager class.

* Optimized import csvs

used new db_manager importing functions to complete the import csv headers.
started working in import ratings from csv.
fixed a formatting issue in db_manager - missing a new line in between functions and it was bothering me.

* Continued working on importing csv

Continued working on importing csv. The names work but the ratings do not import smoothly.

* finished testing excel import button simple pass

* Progress on csv import

progress update

* FIXED CSV IMPORT

Finally fixed the csv import!
Now imports from a template file like TestNewScenarioImport.csv

* Pruning Operation Printin

Removed some unnecessary print lines to have a stable state before
continuing to optimize other operations.

* Optimized import csv operations

Used chatgpt to find an optimization I was happy with.

* rudimentary start to export_csvs

* fixed export to csv functionality

---------

Co-authored-by: pm8k <astromars42@gmail.com>
Co-authored-by: pm8k <matthew.russell@getcatalyst.io>