Added UI improvements
Added control tab to switch between tabs in UI
Added ESC button control to db_load_ui to facilitate testing
• (it should fall back on using the default DB but not sure about this on remote PCs)
Rearranged elements of the UI so that the team names and scenarios are displayed accross the entire application.
Stacked all buttons and checkboxes on the left side of the UI for 2nd tab.