Added "Avoid poor matchups"
This is a new way to score the values of each tree and to allow the trees to be sorted according to that method.
This will check which the lower of the two values is and it will add that value up the tree. This effectively
calculates the lowest possible value of each branch. So the branches with the highest values
shold include the least number of low value pairings.