Pairing algorithm (#16)
* Changes to be committed:
	modified:   qtr_pairing_process/tree_generator.py
	modified:   qtr_pairing_process/ui_manager.py

enhanced the optimize button. It's now toggling between sort and unsort functionality.

* Added "Avoid poor matchups"

This is a new way to score the values of each tree and to allow the trees to be sorted according to that method.
This will check which the lower of the two values is and it will add that value up the tree. This effectively
calculates the lowest possible value of each branch. So the branches with the highest values
shold include the least number of low value pairings.

* added error handling when unsorting tree

Unsort would break if it couldn't find a node to remove it. I've added protections against this.

* advanced sorting under construction

* minor updates in progress

* Adding UI updates and copyright declarations

Added ui update framework.
TODO refactor the ui code so that we can separate out the csv functions.
ui code is getting too cluttered.

* info_grid updates

added functions to handle the first two columns in the info grid.
get score button will call all info grid functions currently.

* Finished inital analysis columns

Originally planned cols for analysis finished.
Collected info on the next set of info to present:

SUM COLUMNS AND COMPARE TO FLOOR
• try to recommend or highlight dropping the player with the highest margin???

* committing additional notes recorded in txt file

* added check_margins

Added a new operation to display data in the info grid.
Not confident this adds value and maybe just junk data.

* optimized and trimmed new operations

* converted data for new opersations

changed a column to max/min margin
added column to display sum of all margins